fix #723: Ending a drag with a right-click results in panels incorrectly swallowing mouse events
