+++ lib/global/event-handlers/onDocumentContextMenu.ts
@@ -1,0 +1,9 @@
+import { completeResize } from "../utils/completeResize.ts";
+
+export function onDocumentContextMenu(event: MouseEvent) {
+  if (event.defaultPrevented) {
+    return;
+  }
+
+  completeResize(event, false);
+}
