+++ lib/global/event-handlers/onDocumentPointerUp.ts
@@ -1,12 +1,4 @@
-import { updateCursorStyle } from "../cursor/updateCursorStyle";
-import {
-  getMountedGroupState,
-  updateMountedGroup
-} from "../mutable-state/groups";
-import {
-  getInteractionState,
-  updateInteractionState
-} from "../mutable-state/interactions";
+import { completeResize } from "../utils/completeResize.ts";
 
 export function onDocumentPointerUp(event: PointerEvent) {
   if (event.defaultPrevented) {
@@ -15,31 +7,5 @@ export function onDocumentPointerUp(event: PointerEvent) {
     return;
   }
 
-  const interactionState = getInteractionState();
-
-  switch (interactionState.state) {
-    case "active": {
-      updateInteractionState({
-        cursorFlags: 0,
-        state: "inactive"
-      });
-
-      if (interactionState.hitRegions.length > 0) {
-        updateCursorStyle(event.currentTarget as Document);
-
-        // Dispatch one more "change" event after the interaction state has been reset.
-        // Groups use this as a signal to call onLayoutChanged.
-        // This is the canonical user-pointer-up site, so flag the dispatch with
-        // isUserInteraction: true. See #716.
-        interactionState.hitRegions.forEach((hitRegion) => {
-          const groupState = getMountedGroupState(hitRegion.group.id, true);
-          updateMountedGroup(hitRegion.group, groupState, {
-            isUserInteraction: true
-          });
-        });
-
-        event.preventDefault();
-      }
-    }
-  }
+  completeResize(event, true);
 }
